Modificación de herencia de 'Recurso' a herencia multitabla
