+++ app_reservas/models/Aula.py
@@ -1,10 +1,7 @@
 # coding=utf-8
 
-import json
-
 from django.db import models
 
-from app_reservas.adapters.google_calendar import obtener_eventos
 from .Recurso import Recurso
 
 
@@ -30,30 +27,8 @@ class Aula(Recurso):
             nombre_corto = self.nombre
         return nombre_corto
 
-    def get_eventos(self):
-        return obtener_eventos(self.calendar_codigo)
-
-    def get_eventos_json(self):
-        eventos = self.get_eventos()
-        eventos_json = '['
-        primera_iteracion = True
-        for evento in eventos:
-            if primera_iteracion:
-                primera_iteracion = False
-            else:
-                eventos_json += ','
-            evento_str = json.dumps({
-                'title': evento['titulo'],
-                'start': evento['inicio_str'],
-                'end': evento['fin_str'],
-                'resourceId': str(self.id)
-            })
-            eventos_json += evento_str
-        eventos_json += ']'
-        return eventos_json
-
     # Información de la clase
-    class Meta(Recurso.Meta):
+    class Meta:
         app_label = 'app_reservas'
         verbose_name = 'Aula'
         verbose_name_plural = 'Aulas'
